Update README.md
The series is set in the fictional continents of Westeros and Essos and follows the power struggles among noble families as they vie for control of the Iron Throne and the Seven Kingdoms of Westeros. It's renowned for its complex characters, intricate plotlines, political intrigue, and themes of power, loyalty, and the consequences of one's actions.